refactor cli utils
